testproject/requirements.txt
# baseapp dependencies
djangorestframework==3.16.1
djangorestframework-simplejwt[crypto]==5.5.1
drf-nested-routers==0.93.5
easy_thumbnails==2.10.1
rest-social-auth==8.4.0

# test
freezegun==1.2.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-asyncio==0.21.1
pytest-celery==1.2.1

responses==0.25.3
urllib3==2.6.3
# test factories
factory-boy==3.2.1

# lint
black==24.3.0
flake8==7.0.0
isort==5.13.2

# channels
channels==4.0.0
channels_redis==4.0.0

django-notifications-hq @ git+https://github.com/openwisp/django-notifications.git@f4918c2409a734ec1aa9798d86c3a729b505f753
django-push-notifications[wp,apns-async,fcm]==3.3.0

-e ./[auth,referrals,graphql,pages,payments,notifications,urlshortening,messagetemplates,wagtail,socialauth,pdf]
